improved linkdb make targets

+++ Makefile
@@ -74,10 +74,11 @@ ALL_MONO_IDSDONE   := $(patsubst ${INDEX_TMPDIR}/%.dedup,done/%.ids.done,${ALL_M
 ##------------------------
 
 
-ALL_ALG_URLS := $(sort $(patsubst %,https:%,$(shell find ${OPUSRELEASE}/ -name statistics.yaml | \
+ALL_ALG_URLS  := $(sort $(patsubst %,https:%,$(shell find ${OPUSRELEASE}/ -name statistics.yaml | \
 						xargs grep 'xml/${LANGPAIR}.xml.gz' | cut -f4 -d:)))
-ALL_ALG_DONE := $(patsubst ${STORAGE_BASE}%.xml.gz,done/%.done,${ALL_ALG_URLS})
-ALL_LINK_DBS := $(subst /xml/,/,$(patsubst done/%/${LANGPAIR}.done,sqlite/%/${LINKDB_LANGPAIR}.db,${ALL_ALG_DONE}))
+ALL_ALG_DONE  := $(patsubst ${STORAGE_BASE}%.xml.gz,done/%.done,${ALL_ALG_URLS})
+ALL_LINK_DONE := $(patsubst ${STORAGE_BASE}%.xml.gz,done/%.linkdb.done,${ALL_ALG_URLS})
+ALL_LINK_DBS  := $(subst /xml/,/,$(patsubst done/%/${LANGPAIR}.done,sqlite/%/${LINKDB_LANGPAIR}.db,${ALL_ALG_DONE}))
 
 
 ## alignment databases
@@ -86,8 +87,10 @@ ALL_LINK_DBS := $(subst /xml/,/,$(patsubst done/%/${LANGPAIR}.done,sqlite/%/${LI
 ##   LATEST_LINK_DB        = links from latest releases (pointing to sentence index IDs)
 ##   LATEST_LINK_DB_MERGED = flags marking that the latest release has been merged
 
-LINK_DB                := ${LANGPAIR3}.db
+LINK_DB               := ${LANGPAIR3}.db
 LATEST_LINK_DB        := sqlite/${LANGPAIR3}.db
+OLDER_LINK_DB         := sqlite/${LANGPAIR3}.releases.db
+
 LINK_DB_MERGED        := $(patsubst %.db,%.merged,${ALL_LINK_DBS})
 LATEST_LINK_DB_MERGED := $(sort $(shell echo "${LINK_DB_MERGED}" | tr ' ' "\n" | cut -f1,2,4 -d/))
 
@@ -156,7 +159,6 @@ all-links:
 
 .PHONY: linkdb
 linkdb: ${LATEST_LINK_DB}
-
 
 
 
@@ -380,6 +382,7 @@ ${ALL_ALG_DONE}: ${INDEX_TMPDIR}/${LINK_DB}
 
 
 
+
 ##--------------------------------------------------------------------------------
 ## database of linked source and target sentences
 ##  --> maps internal sentence IDs to internal link IDs
@@ -404,7 +407,7 @@ ${LATEST_LINK_DB}: ${LINK_DB} ${ALL_LINK_DBS}
 	sqlite3 $< ".dump bitexts" | ${MODIFY_DB_DUMP} | sqlite3 $@
 	echo "${CREATE_UNIQUE_INDEX} idx_bitexts ON bitexts (corpus,version,fromDoc,toDoc)" | sqlite3 $@
 	${SCRIPTDIR}add_bitext_range.py $@
-	${SCRIPTDIR}add_corpus_range.py $@
+	${SCRIPTDIR}add_corpus_range.py -d $@
 
 
 ## phony target to merge link tables from each corpus
@@ -425,12 +428,14 @@ merge-latest-linkdbs: ${LATEST_LINK_DB_MERGED}
 ## --> pre-requisite databases will be copied to a temporary location
 ## --> this makes lookup much faster (assuming that the tmpdisk is a fast local disk)
 
-LINKDB_PREREQUISITES := ${INDEX_TMPDIR}/linkdb/${LINK_DB} \
-			${INDEX_TMPDIR}/linkdb/${SRCLANG_IDX_DB} \
-			${INDEX_TMPDIR}/linkdb/${TRGLANG_IDX_DB}
+LINKDB_PREREQ_TMPDIR := ${INDEX_TMPDIR}/linkdb
+
+LINKDB_PREREQUISITES := ${LINKDB_PREREQ_TMPDIR}/${LINK_DB} \
+			${LINKDB_PREREQ_TMPDIR}/${SRCLANG_IDX_DB} \
+			${LINKDB_PREREQ_TMPDIR}/${TRGLANG_IDX_DB}
 
 .INTERMEDIATE: ${LINKDB_PREREQUISITES}
-${LINKDB_PREREQUISITES}: ${INDEX_TMPDIR}/linkdb/%: %
+${LINKDB_PREREQUISITES}: ${LINKDB_PREREQ_TMPDIR}/%: %
 	mkdir -p $(dir $@)
 	rsync -av $< $@
 
@@ -448,21 +453,32 @@ ${ALL_LINK_DBS}: ${LINKDB_PREREQUISITES}
 
 
 
+
+
+
+
+
+
+
+
+
 ## initialize the global link database in local tmp dir with fast I/O
 ## declare this to be an intermediate file to remove it after finishing the process
 
-TMP_LINK_DB := ${INDEX_TMPDIR}/${LATEST_LINK_DB}
-.INTERMEDIATE: ${TMP_LINK_DB}
+TMP_LATEST_LINK_DB := ${INDEX_TMPDIR}/${LATEST_LINK_DB}
+TMP_OLDER_LINK_DB  := ${INDEX_TMPDIR}/${OLDER_LINK_DB}
+TMP_LINK_DB        := ${TMP_LATEST_LINK_DB}
+.INTERMEDIATE: ${TMP_LINK_DB} ${TMP_LATEST_LINK_DB} ${TMP_OLDER_LINK_DB}
 
 
 ## open with timeout to allow concurrent access
 ## but that still does not seem to work well (skip timeout?)
 SQLITE3 = sqlite3 -cmd ".timeout 100000"
 
-${TMP_LINK_DB}:
-	$(call retrieve,${LATEST_LINK_DB})
+${TMP_LATEST_LINK_DB} ${TMP_OLDER_LINK_DB}:
+	$(call retrieve,$(patsubst ${INDEX_TMPDIR}/%,%,$@))
 	mkdir -p $(dir $@)
-	if [ -e ${LATEST_LINK_DB} ]; then rsync -av ${LATEST_LINK_DB} $@; fi
+	if [ -e $(patsubst ${INDEX_TMPDIR}/%,%,$@) ]; then rsync -av $(patsubst ${INDEX_TMPDIR}/%,%,$@) $@; fi
 	@echo "${CREATE_TABLE} linkedsource (sentID INTEGER,linkID INTEGER,bitextID INTEGER,PRIMARY KEY(linkID,sentID) )" | ${SQLITE3} $@
 	@echo "${CREATE_TABLE} linkedtarget (sentID INTEGER,linkID INTEGER,bitextID INTEGER,PRIMARY KEY(linkID,sentID) )" | ${SQLITE3} $@
 	@echo "${CREATE_INDEX} idx_linkedsource_bitext ON linkedsource (bitextID,sentID)" | ${SQLITE3} $@
@@ -479,8 +495,9 @@ ${TMP_LINK_DB}:
 	@echo "${CREATE_INDEX} idx_bitextid ON links (bitextID)" | ${SQLITE3} $@
 	@echo "${CREATE_TABLE} corpora (corpus TEXT,version TEXT,srclang TEXT,trglang TEXT,srclang3 TEXT,trglang3 TEXT)" | sqlite3 $@
 	@echo "${CREATE_UNIQUE_INDEX} idx_corpora ON corpora (corpus,version,srclang,trglang,srclang3,trglang3)" | sqlite3 $@
+	@sqlite3 ${LINK_DB} ".dump bitexts" | ${MODIFY_DB_DUMP} | sqlite3 $@
+	@echo "${CREATE_UNIQUE_INDEX} idx_bitexts ON bitexts (corpus,version,fromDoc,toDoc)" | sqlite3 $@
 	@echo "PRAGMA journal_mode=WAL" | ${SQLITE3} $@
-
 
 
 ## merge links into the global link database
@@ -514,8 +531,11 @@ ${LATEST_LINK_DB_MERGED}: ${TMP_LINK_DB}
 	  if [ ! -e sqlite/$$c/$$l/${LINKDB_LANGPAIR}.merged ]; then \
 	    echo "add links for $$c/$$l/${LINKDB_LANGPAIR}"; \
 	    if [ ! -e sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db ]; then \
-		${MAKE} sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db || \
-		echo "cannot make sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db"; \
+	      if [ -e ${OPUSRELEASE}/$$c/$$l/statistics.yaml ]; then \
+	        if [ `grep '${LINKDB_LANGPAIR}.xml.gz' ${OPUSRELEASE}/$$c/$$l/statistics.yaml | wc -l` -gt 0 ]; then \
+	          ${MAKE} sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db; \
+	        fi \
+	      fi \
 	    fi; \
 	    if [ -e sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db ]; then \
 	      rsync sqlite/$$c/$$l/${LINKDB_LANGPAIR}.db ${INDEX_TMPDIR}/$$c-$$l-${LINKDB_LANGPAIR}.db; \
@@ -573,6 +593,174 @@ ${LATEST_LINK_DB_MERGED}: ${TMP_LINK_DB}
 ##-----------------------------------------------------------------------------
 
 
+
+new-linkdb: ${ALL_LINK_DONE}
+	${SCRIPTDIR}add_bitext_range.py ${LATEST_LINK_DB}
+	${SCRIPTDIR}add_bitext_range.py ${OLDER_LINK_DB}
+
+new-linkdb-local:
+	@$(call lockfile,${LATEST_LINK_DB})
+	@$(call lockfile,${OLDER_LINK_DB})
+	@$(call create_linkdb,${LATEST_LINK_DB})
+	@$(call create_linkdb,${OLDER_LINK_DB})
+	@echo "make link DBs in ${INDEX_TMPDIR}"
+	@${MAKE} -j 1 \
+		LATEST_LINK_DB=${INDEX_TMPDIR}/${LATEST_LINK_DB} \
+		OLDER_LINK_DB=${INDEX_TMPDIR}/${OLDER_LINK_DB} \
+		new-linkdb
+	mv -f ${INDEX_TMPDIR}/${LATEST_LINK_DB} ${LATEST_LINK_DB}
+	mv -f ${INDEX_TMPDIR}/${OLDER_LINK_DB} ${OLDER_LINK_DB}
+	@$(call unlockfile,${LATEST_LINK_DB})
+	@$(call unlockfile,${OLDER_LINK_DB})
+
+
+redo-langs.txt:
+	find done -name '*-*.done' | cut -f5 -d/ | cut -f1 -d. | sort -u > $@
+
+REDO_LANGS = $(patsubst %,%.done,${shell cat redo-langs.txt})
+
+redo-langs: ${REDO_LANGS}
+
+${REDO_LANGS}:
+	-${MAKE} LANGPAIR=$(@:.done=) LINKDB_PREREQUISITES='' new-linkdb-local
+
+
+
+
+## extract linkdb data for a given corpus release
+## merge all data into the central link DBs
+##   - LATEST_LINK_DB for the latest corpus releases
+##   - OLDER_LINK_DB for all other corpus releases
+
+
+## SQL comands to create a link databass
+
+CREATE_LINKDB := \
+	${CREATE_TABLE} linkedsource (sentID INTEGER,linkID INTEGER,bitextID INTEGER,PRIMARY KEY(linkID,sentID) ); \
+	${CREATE_TABLE} linkedtarget (sentID INTEGER,linkID INTEGER,bitextID INTEGER,PRIMARY KEY(linkID,sentID) ); \
+	${CREATE_INDEX} idx_linkedsource_bitext ON linkedsource (bitextID,sentID) ; \
+	${CREATE_INDEX} idx_linkedtarget_bitext ON linkedtarget (bitextID,sentID) ;\
+	${CREATE_INDEX} idx_linkedsource_linkid ON linkedsource (linkID); \
+	${CREATE_INDEX} idx_linkedtarget_linkid ON linkedtarget (linkID); \
+	${CREATE_INDEX} idx_linkedsource_sentid ON linkedsource (sentID); \
+	${CREATE_INDEX} idx_linkedtarget_sentid ON linkedtarget (sentID); \
+	${CREATE_TABLE} links (linkID INTEGER NOT NULL PRIMARY KEY,bitextID, \
+                                      srcIDs TEXT,trgIDs TEXT,srcSentIDs TEXT,trgSentIDs TEXT, \
+                                      alignType TEXT,alignerScore REAL,cleanerScore REAL); \
+	${CREATE_UNIQUE_INDEX} idx_links ON links (bitextID,srcIDs,trgIDs); \
+	${CREATE_INDEX} idx_aligntype ON links (bitextID,alignType); \
+	${CREATE_INDEX} idx_bitextid ON links (bitextID); \
+	${CREATE_TABLE} corpora (corpus TEXT,version TEXT,srclang TEXT,trglang TEXT,srclang3 TEXT,trglang3 TEXT); \
+	${CREATE_UNIQUE_INDEX} idx_corpora ON corpora (corpus,version,srclang,trglang,srclang3,trglang3); \
+	${CREATE_TABLE} bitexts (corpus TEXT,version TEXT,fromDoc TEXT,toDoc TEXT); \
+	${CREATE_UNIQUE_INDEX} idx_bitexts ON bitexts (corpus,version,fromDoc,toDoc); \
+	ATTACH DATABASE '${LINK_DB}' as l; \
+	${INSERT_INTO} bitexts SELECT * FROM l.bitexts ORDER BY rowid; \
+	PRAGMA journal_mode=WAL;
+
+
+## additional helper functions and variables
+
+create_linkdb =	if [ ! -s $1 ]; then echo "create $1";mkdir -p $(dir $1); echo "${CREATE_LINKDB}" | ${SQLITE3} $1; fi
+
+LINKDB_MATCH_CORPUS      := corpus='$$c' AND srclang='${SRCLANG}' AND trglang='${TRGLANG}'
+LINKDB_MATCH_CORPUS_DOCS := corpus='$$c' AND fromDoc LIKE '${SRCLANG}/%' AND toDoc LIKE '${TRGLANG}/%'
+LINKDB_CORPUS_SELECT     := SELECT DISTINCT rowid FROM bitexts WHERE ${LINKDB_MATCH_CORPUS_DOCS}
+LINKDB_MATCH_BITEXT      := bitextID IN ( ${LINKDB_CORPUS_SELECT} )
+LINKDB_COUNT_RELEASES    := echo "SELECT rowid FROM corpora WHERE ${LINKDB_MATCH_CORPUS}" \
+				| ${SQLITE3} ${LATEST_LINK_DB} | wc -l
+
+
+${ALL_LINK_DONE}: ${LINKDB_PREREQUISITES}
+
+## create temporary database for the selected corpus/release/langpair
+
+	@mkdir -p $(dir ${INDEX_TMPDIR}/$@) $(dir ${LATEST_LINK_DB}) $(dir ${OLDER_LINK_DB})
+	@if [ -e sqlite/$(word 2,$(subst /, ,$@))/$(word 3,$(subst /, ,$@))/$(notdir $(@:.linkdb.done=.db)) ]; then \
+	  echo "copy sqlite DB to ${INDEX_TMPDIR}/$@"; \
+	  cp sqlite/$(word 2,$(subst /, ,$@))/$(word 3,$(subst /, ,$@))/$(notdir $(@:.linkdb.done=.db)) ${INDEX_TMPDIR}/$@; \
+	elif [ -e sqlite-old/$(word 2,$(subst /, ,$@))/$(word 3,$(subst /, ,$@))/$(notdir $(@:.linkdb.done=.db)) ]; then \
+	  echo "copy sqlite-old DB to ${INDEX_TMPDIR}/$@"; \
+	  cp sqlite-old/$(word 2,$(subst /, ,$@))/$(word 3,$(subst /, ,$@))/$(notdir $(@:.linkdb.done=.db)) ${INDEX_TMPDIR}/$@; \
+	else \
+	  if [ "${LINKDB_PREREQUISITES}" != "" ]; then \
+	    ${LINK2SQLITE} $^ ${INDEX_TMPDIR}/$@ $(word 2,$(subst /, ,$@)) $(word 3,$(subst /, ,$@)); \
+	  else \
+	    if [ -e ${TMPDIR}/${SRCLANG_IDX_DB} ]; then S=${TMPDIR}/${SRCLANG_IDX_DB}; else S=${SRCLANG_IDX_DB}; fi; \
+	    if [ -e ${TMPDIR}/${TRGLANG_IDX_DB} ]; then T=${TMPDIR}/${TRGLANG_IDX_DB}; else T=${TRGLANG_IDX_DB}; fi; \
+	    echo "${LINK2SQLITE} ${LINK_DB} $$S $$T ${INDEX_TMPDIR}/$@ $(word 2,$(subst /, ,$@)) $(word 3,$(subst /, ,$@))"; \
+	    mkdir -p $(dir ${INDEX_TMPDIR}/$@); \
+	    ${LINK2SQLITE} ${LINK_DB} $$S $$T ${INDEX_TMPDIR}/$@ $(word 2,$(subst /, ,$@)) $(word 3,$(subst /, ,$@)); \
+	  fi \
+	fi
+
+## if the release is the latest one for that corpus/langpair
+## --> move other release data (if they exist) from latest link DB to DB of other releases
+## --> add links from temporary DB to latest link DB
+## otherwise:
+## --> add links from temporary DB to DB of other releases
+## do all this in tempdir (which requires locking the DB, creating, syncing back)
+
+	@if [ -s ${INDEX_TMPDIR}/$@ ]; then \
+	  c=$(word 2,$(subst /, ,$@)); \
+	  v=$(word 3,$(subst /, ,$@)); \
+	  l=`grep 'latest_release:' ${OPUSRELEASE}/$$c/info.yaml | cut -f2 -d' ' | xargs`; \
+	  if [ ! -e ${OPUSRELEASE}/$$c/$$l/statistics.yaml ] || \
+	     [ `grep '$(notdir $(@:.linkdb.done=.xml.gz))' ${OPUSRELEASE}/$$c/$$l/statistics.yaml | wc -l` -eq 0 ]; then \
+	     l=`grep '$(notdir $(@:.linkdb.done=.xml.gz))' ${OPUSRELEASE}/$$c/index.txt | tail -1 | cut -f1 -d/`; \
+	  fi; \
+	  if [ "$$v" == "$$l" ]; then \
+	    echo "release $$v is the latest release for $$c"; \
+	    $(call lockfile,${LATEST_LINK_DB}); \
+	    $(call create_linkdb,${LATEST_LINK_DB}); \
+	    if [ `${LINKDB_COUNT_RELEASES}` -gt 0 ]; then \
+	      echo "found previous release(s) in ${LATEST_LINK_DB}"; \
+	      echo "move links for $$c/*/${LINKDB_LANGPAIR} to DB of older releases"; \
+	      $(call lockfile,${OLDER_LINK_DB}); \
+	      $(call create_linkdb,${OLDER_LINK_DB}); \
+	      echo "ATTACH DATABASE '${LATEST_LINK_DB}' as l; \
+	            ${INSERT_INTO} corpora SELECT * from l.corpora WHERE ${LINKDB_MATCH_CORPUS}; \
+		    ${INSERT_INTO} links SELECT * FROM l.links WHERE ${LINKDB_MATCH_BITEXT} ; \
+		    ${INSERT_INTO} linkedsource SELECT * FROM l.linkedsource WHERE ${LINKDB_MATCH_BITEXT}; \
+		    ${INSERT_INTO} linkedtarget SELECT * FROM l.linkedtarget WHERE ${LINKDB_MATCH_BITEXT};" \
+	      | ${SQLITE3} ${OLDER_LINK_DB}; \
+	      ${SCRIPTDIR}add_corpus_range.py -d ${OLDER_LINK_DB} -c $$c -s ${SRCLANG} -t ${TRGLANG}; \
+	      $(call unlockfile,${OLDER_LINK_DB}); \
+	      echo "remove links for $$c/*/${LINKDB_LANGPAIR} from link DB of latest releases"; \
+	      echo "DELETE FROM corpora WHERE ${LINKDB_MATCH_CORPUS}; \
+	            DELETE FROM links WHERE ${LINKDB_MATCH_BITEXT}; \
+		    DELETE FROM linkedsource WHERE ${LINKDB_MATCH_BITEXT}; \
+		    DELETE FROM linkedtarget WHERE ${LINKDB_MATCH_BITEXT};" \
+	      | ${SQLITE3} ${LATEST_LINK_DB}; \
+	    fi; \
+	    echo "add links for $$c/$$v/${LINKDB_LANGPAIR} to link DB of latest releases"; \
+	    echo "ATTACH DATABASE '${INDEX_TMPDIR}/$@' as l; \
+	          ${INSERT_INTO} corpora VALUES('$$c','$$l','${SRCLANG}','${TRGLANG}','${SRCLANG3}','${TRGLANG3}'); \
+		  ${INSERT_INTO} links SELECT * FROM l.links; \
+		  ${INSERT_INTO} linkedsource SELECT * FROM l.linkedsource; \
+		  ${INSERT_INTO} linkedtarget SELECT * FROM l.linkedtarget;" \
+	    | ${SQLITE3} ${LATEST_LINK_DB}; \
+	    rm -f ${INDEX_TMPDIR}/$@; \
+	    ${SCRIPTDIR}add_corpus_range.py -d ${LATEST_LINK_DB} -c $$c -v $$v -s ${SRCLANG} -t ${TRGLANG}; \
+	    $(call unlockfile,${LATEST_LINK_DB}); \
+	  else \
+	    echo "release $$v is an older release for $$c"; \
+	    $(call lockfile,${OLDER_LINK_DB}); \
+	    $(call create_linkdb,${OLDER_LINK_DB}); \
+	    echo "add links for $$c/$$v/${LINKDB_LANGPAIR} to link DB of older releases"; \
+	    echo "ATTACH DATABASE '${INDEX_TMPDIR}/$@' as l; \
+	          ${INSERT_INTO} corpora VALUES('$$c','$$v','${SRCLANG}','${TRGLANG}','${SRCLANG3}','${TRGLANG3}'); \
+		  ${INSERT_INTO} links SELECT * FROM l.links; \
+		  ${INSERT_INTO} linkedsource SELECT * FROM l.linkedsource; \
+		  ${INSERT_INTO} linkedtarget SELECT * FROM l.linkedtarget;" \
+	    | ${SQLITE3} ${OLDER_LINK_DB}; \
+	    rm -f ${INDEX_TMPDIR}/$@; \
+	    ${SCRIPTDIR}add_corpus_range.py -d ${OLDER_LINK_DB} -c $$c -v $$v -s ${SRCLANG} -t ${TRGLANG}; \
+	    $(call unlockfile,${OLDER_LINK_DB}); \
+	  fi; \
+	fi
+	@mkdir -p $(dir $@)
+	@touch $@
 
 
 
@@ -671,7 +859,7 @@ add-corpora2linkdb: $(CORPORA_LINK_DBS)
 	                                 WHERE fromDoc LIKE '$$s/%' AND toDoc LIKE '$$t/%';" \
 	    | sqlite3 $<; \
 	  done )
-	${SCRIPTDIR}add_corpus_range.py $<
+	${SCRIPTDIR}add_corpus_range.py -d $<
 
 
 ##--------------------------------------------------
@@ -686,12 +874,12 @@ remove-linkdb-overlaps:
 	${MAKE} RERUN_LINK_EXTRACTION=1 check-linkdb-overlaps
 
 %.check.db: %.db
-	if [ ! -s $< ]; then \
+	@if [ ! -s $< ]; then \
 	  if [ -s $<.backup ]; then mv -f $<.backup $<; fi; \
 	  if [ -s sqlite/$<.backup ]; then mv -f sqlite/$<.backup sqlite/$<; fi; \
 	fi
-	${MAKE} $(<:.db=.corpora.db)
-	if [ `scripts/check_corpus_ranges.py $< | wc -l` -gt 0 ]; then \
+	@if [ "${RERUN_LINK_EXTRACTION}" == "1" ]; then ${MAKE} $(<:.db=.corpora.db); fi
+	@if [ `scripts/check_corpus_ranges.py $< | wc -l` -gt 0 ]; then \
 	  echo "# overlaps in $<"; \
 	  P=`scripts/check_corpus_ranges.py $< | cut -f3,4 -d, | cut -f1 -d' ' | sort -u | tr ',' '-'`;\
 	  if [ "${RERUN_LINK_EXTRACTION}" == "1" ]; then \
@@ -864,6 +1052,32 @@ ${INDEX_TMPDIR}/%.jsonl:
 		-l ${LANGUAGE} > $@
 
 
+.PHONY: opensub-docs
+opensub-docs:
+	${MAKE} CORPUS=OpenSubtitles VERSION=v2018 LANGUAGE=en corpus-docs
+	${MAKE} CORPUS=OpenSubtitles VERSION=v2018 LANGUAGE=en corpus-docs
+
+.PHONY: europarl-docs
+europarl-docs:
+	${MAKE} CORPUS=Europarl VERSION=v8 LANGUAGE=en corpus-docs
+	${MAKE} CORPUS=Europarl VERSION=v8 LANGUAGE=en corpus-docs
+
+.PHONY: unpc-docs
+unpc-docs:
+	${MAKE} CORPUS=UNPC VERSION=v1.0 LANGUAGE=en corpus-docs
+	${MAKE} CORPUS=UNPC VERSION=v1.0 LANGUAGE=en corpus-docs
+
+.PHONY: corpus-docs
+corpus-docs: ${CORPUS}_${VERSION}.${LANGUAGE}.jsonl.gz ${CORPUS}_${VERSION}.${LANGUAGE}.txt.gz
+
+${CORPUS}_${VERSION}.${LANGUAGE}.jsonl.gz:
+	${SCRIPTDIR}/opus_get_documents.py -c ${CORPUS} -r ${VERSION} -l ${LANGUAGE} -j | ${GZIP} -c > $@
+
+${CORPUS}_${VERSION}.${LANGUAGE}.txt.gz:
+	${SCRIPTDIR}/opus_get_documents.py -c ${CORPUS} -r ${VERSION} -l ${LANGUAGE} | ${GZIP} -c > $@
+
+
+
 ## unicode cleanup
 
 # FIX_UNICODE := perl -CS -pe 'tr[\x{9}\x{A}\x{D}\x{20}-\x{D7FF}\x{E000}-\x{FFFD}\x{10000}-\x{10FFFF}][]cd;'
@@ -921,7 +1135,18 @@ else
 retrieve = echo "downloading files is disabled (skip retrieving $1)"
 endif
 
+lockfile = ( while [ -e ${1}.lock ]; do \
+	       echo "waiting for exclusive access to ${1}"; \
+	       sleep 10; \
+	       find $(dir ${1}) -name $(notdir ${1}).lock -mtime +1 -delete; \
+	     done; \
+	     touch ${1}.lock; )
 
+unlockfile = rm -f ${1}.lock
+
+
+test-lock:
+	@$(call lockfile,test)
 
 ## OLD: retrieve files from storage as a PHONY make target
 ## --> this is slow because of a separate make call
